backend/requirements.txt
# Main dependencies for Claude chat app
fastapi>=0.109.1
uvicorn>=0.27.0
websockets>=12.0
python-multipart>=0.0.6

# API client for Anthropic Claude
anthropic>=0.18.0

# Utilities
Pillow>=10.2.0
python-dotenv>=1.0.0
httpx>=0.26.0
python-docx>=0.8.11
pytesseract>=0.3.10